+++ src/main/java/io/github/soydivision/book/BooksRepository.java
@@ -2,6 +2,6 @@ package io.github.soydivision.book;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BooksRepository extends CrudRepository<Book, Integer>
+public interface BooksRepository extends CrudRepository<Book, Long>
 {
 }
